+++ lib/anyway/ext/hash.rb
@@ -21,9 +21,10 @@ module Anyway
 
           last_key = path.pop
           hash = path.reduce(self) do |hash, k|
-            hash[k] = {} unless hash.key?(k)
+            hash[k] = {} unless hash.key?(k) && hash[k].is_a?(::Hash)
             hash[k]
           end
+
           hash[last_key] = val
         end